fix: ensure nodeadm config exists before nodeadm upgrade
