+++ pkg/provider/testdata/ssm-custom-agent/expected.yaml
@@ -141,6 +141,94 @@ Stages:
     UnpackImages: null
     Users: null
   - After: null
+    Commands: null
+    DataSources:
+      Path: ""
+      Providers: null
+    DeleteEntities: null
+    Directories: null
+    Dns:
+      DnsOptions: null
+      DnsSearch: null
+      Nameservers: null
+      Path: ""
+    Downloads: null
+    EnsureEntities: null
+    Environment: null
+    EnvironmentFile: ""
+    Files:
+    - Content: |
+        apiVersion: node.eks.aws/v1alpha1
+        kind: NodeConfig
+        metadata:
+          creationTimestamp: null
+        spec:
+          cluster:
+            name: eks-hybrid-cluster
+            region: us-east-1
+          containerd:
+            config: |
+              [plugins."io.containerd.grpc.v1.cri".containerd]
+              discard_unpacked_layers = false
+          hybrid:
+            ssm:
+              activationCode: 12345678-1234-1234-1234-123456789012
+              activationId: 12345678-1234-1234-1234-123456789012
+          instance:
+            localStorage: {}
+          kubelet:
+            config:
+              shutdownGracePeriod: 30s
+            flags:
+            - --node-labels=abc.company.com/test-label=true
+      Encoding: ""
+      Group: 0
+      Owner: 0
+      OwnerString: ""
+      Path: /opt/nodeadmutil/node-config.yaml
+      Permissions: 416
+    Git:
+      Auth:
+        Insecure: false
+        Password: ""
+        PrivateKey: ""
+        PublicKey: ""
+        Username: ""
+      Branch: ""
+      BranchOnly: false
+      Path: ""
+      URL: ""
+    Hostname: ""
+    If: ""
+    Layout:
+      Device: null
+      Expand: null
+      Parts: null
+    Modules: null
+    Name: Generate nodeadm config file
+    Node: ""
+    OnlyIfArch: ""
+    OnlyIfOs: ""
+    OnlyIfOsVersion: ""
+    OnlyIfServiceManager: ""
+    Packages:
+      Install: null
+      Refresh: false
+      Remove: null
+      Upgrade: false
+    SSHKeys: null
+    Sysctl: null
+    Systemctl:
+      Disable: null
+      Enable: null
+      Mask: null
+      Overrides: null
+      Start: null
+    SystemdFirstBoot: null
+    TimeSyncd: null
+    UnpackImages: null
+    Users: null
+  - After: null
     Commands:
     - 'bash /opt/nodeadmutil/scripts/nodeadm-install.sh 1.30.0 ssm /opt/nodeadmutil
       false '
@@ -239,94 +327,6 @@ Stages:
       Parts: null
     Modules: null
     Name: Run Nodeadm Upgrade
-    Node: ""
-    OnlyIfArch: ""
-    OnlyIfOs: ""
-    OnlyIfOsVersion: ""
-    OnlyIfServiceManager: ""
-    Packages:
-      Install: null
-      Refresh: false
-      Remove: null
-      Upgrade: false
-    SSHKeys: null
-    Sysctl: null
-    Systemctl:
-      Disable: null
-      Enable: null
-      Mask: null
-      Overrides: null
-      Start: null
-    SystemdFirstBoot: null
-    TimeSyncd: null
-    UnpackImages: null
-    Users: null
-  - After: null
-    Commands: null
-    DataSources:
-      Path: ""
-      Providers: null
-    DeleteEntities: null
-    Directories: null
-    Dns:
-      DnsOptions: null
-      DnsSearch: null
-      Nameservers: null
-      Path: ""
-    Downloads: null
-    EnsureEntities: null
-    Environment: null
-    EnvironmentFile: ""
-    Files:
-    - Content: |
-        apiVersion: node.eks.aws/v1alpha1
-        kind: NodeConfig
-        metadata:
-          creationTimestamp: null
-        spec:
-          cluster:
-            name: eks-hybrid-cluster
-            region: us-east-1
-          containerd:
-            config: |
-              [plugins."io.containerd.grpc.v1.cri".containerd]
-              discard_unpacked_layers = false
-          hybrid:
-            ssm:
-              activationCode: 12345678-1234-1234-1234-123456789012
-              activationId: 12345678-1234-1234-1234-123456789012
-          instance:
-            localStorage: {}
-          kubelet:
-            config:
-              shutdownGracePeriod: 30s
-            flags:
-            - --node-labels=abc.company.com/test-label=true
-      Encoding: ""
-      Group: 0
-      Owner: 0
-      OwnerString: ""
-      Path: /opt/nodeadmutil/node-config.yaml
-      Permissions: 416
-    Git:
-      Auth:
-        Insecure: false
-        Password: ""
-        PrivateKey: ""
-        PublicKey: ""
-        Username: ""
-      Branch: ""
-      BranchOnly: false
-      Path: ""
-      URL: ""
-    Hostname: ""
-    If: ""
-    Layout:
-      Device: null
-      Expand: null
-      Parts: null
-    Modules: null
-    Name: Generate nodeadm config file
     Node: ""
     OnlyIfArch: ""
     OnlyIfOs: ""
